Use default `os_version`
Previously this feedstock needed to move to CentOS 7 for builds before
conda-forge did. So `os_version` was used for that. Since then,
conda-forge moved to GLIBC 2.17 for build compatibility and AlmaLinux 9
for the OS. As a result the `os_version` here has become old, which the
linter warns about. Given this we can drop `os_version` and use the
defaults.